chore(thermal): file renaming
